versions filtering to own file, and fix some logic around snapshots
